+++ src/views/AuthoritiesSearch/AuthoritiesSearch.test.js
@@ -390,7 +390,7 @@ describe('Given AuthoritiesSearch', () => {
   });
 
   describe('when there is a linked record', () => {
-    it('should show number of titles', () => {
+    it('should show number of titles and be a link to the inventory app by authority ID', () => {
       const { getAllByRole } = renderAuthoritiesSearch({
         authorities: authorities.slice(0, 3),
         totalRecords: 3,
@@ -399,6 +399,9 @@ describe('Given AuthoritiesSearch', () => {
       expect(getAllByRole('row')[2].textContent).toContain('Authorized');
       expect(getAllByRole('row')[2].textContent).toContain('Twain, Mark linked');
       expect(getAllByRole('row')[2].textContent).toContain('1');
+      expect(getAllByRole('row')[2].href).toContain(
+        '/inventory?qindex=authorityId&query=5a404f5d-2c46-4426-9f28-db8d26881b31&sort=title'
+      )
     });
   });